refactor: hoist crypto related code to own package
This enables consumers of cryptography related code to share the same
implementations. (We expect vouchers to re-use this code.)
