+++ crypto/signatures.go
@@ -1,4 +1,4 @@
-package state
+package crypto
 
 import (
 	"fmt"
